Add files via upload
GenshinRateUp.java is a monte carlo simulation that runs my previous personal hypothesis for how rates worked with GenshinOutput.txt being the output as an array, Loop.java is a monte carlo sim that runs the actual ruleset that it runs with GenshinRealOutput.txt being the output as an array. 
Image.png is a graph comparing the results of GenshinOutput (blue), GenshinRealOutput (gray), and real data (orange) from real users. Both simulations and the real user data are generated from 1.2 million trials